initial commit
here's a simple dice game (in latvian) which technically one plays as a
drinking game.

That said now there's covid so we don't do that stuff in person anymore,
but why not do this online.

Anyway the game logic itself is in a GenServer state machine that sits
in lib/lucky/game.ex

The logic then is that a supervisor (lib/lucky/game_server.ex), spawns
named games and manages them.

The games get spawned by someone going to a /game/<name_of_the_game>
url.

Then we track the presence of players and when there's at least two of
them, we can start the game.

The ui is pretty neat in that it's completely declarative and based on
the state of the game so we don't have to track any state-transition
logic for it (that all happens in the game logic itself) and we just
send messages to the game and then ask it for what is its state

TODO: make this whole thing use https and umm, clean up the html to
remove the phoenix framework default header