Better deal with people leaving and fix the timer
So the timer now is a live component and thus updates to it no longer
impact the rest of the html (e.g. the bet dropdown) which means we can
show it (and scare people into doing things faster)

Additionally we now deal with everyone leaving (closing tabs or getting
kicked out) and while annoyingly the only way to be sure that noone is
looking at a certain game is to make the game itself send some kind of
timers (where it would kill itself if nobody sends it any events within
a certain timeframe), the UX now behaves more like it should and you
shouldn't see any "ghosts" anymore.

That said I think I need to fix a thing where if you rejoin as the sole
person in a game that was started you probably cannot actually start it
or do anything. _maybe_. But I'd need to check - perhaps it counts as
started and thus you're just there and can play with yourself or
something.